Add ObjectPool with WeakObjectPtr support
ObjectPool contains a std::vector<T> of objects. It keeps track of safeguard values in parallel. The safeguard values are used with WeakObjectPtr to ensure that it references the correct object. If an object is destroyed or replaced the safeguard will no longer match even though the id to the slot is identical.

Objects can be referenced directly using ObjectId which is just an index but it guarantees no access safety or same-object checks.

IsValid functions added to allow validity checks of ids.

WeakObjectPtr is intended to be used with collections (like scenes) to handle user interactions like removal of objects safely.