Make pools register to internal array for simple cleanup
Pools now register themselves on first use to an internal array. The application calls Shutdown that iterates the arrays and triggers the emptying of pools before exiting.

For some reason the registration did not work in the constructor... The array contents reset on startup. Not much to do about it.