Fix viewport aspect ratio bug
Viewports were always rendered with the screen ratio so this bug was not seen until now. The viewport ratio was set to the application ratio.

The camera now gets a ratio value for each time the UBO is updated. It is calculated via the Viewport framebuffer internals.

Also add a status bar.