Fix bug where mouse would not release from viewport
It was possible to drag the mouse fast enough to go outside the viewport and be detected in the imgui section to clear the active viewport. This caused the UI to capture input and the mouse would not get its released event and it would remain invisible.

UpdateActiveViewport now does nothing when the mouse is captured by a viewport.