Change so that raw image data is no longer flipped
Textures must be flipped in OpenGL for UV coordinates to be correct.

We previously flipped images vertically on load to account for this... That is however not a good idea if we are going to involve any image processing at all.

We now flip the image in the shader in the shader instead.

Change the define FLIP_IMAGES_VERTICALLY to undo it.