Load extra 10 points for eyes from mediapipe
The mediapipe facemesh is 468 points. The landmarker task can generate 478 points where the extra ones are 5 per iris. Add extra vertices and triangles to the face when those landmarks are supplied.

Also add ability for the user to change the glBufferVector usage.

Also add F as shortcut for toggle fullscreen.